Compacter doesn't delete parent states for tipsets with height > main branch height
